refactor: entity.Batch -> entity.Batch.Batch
