+++ core/src/main/java/cn/edu/tsinghua/iot/benchmark/client/generate/GenerateBaseClient.java
@@ -21,7 +21,7 @@ package cn.edu.tsinghua.iot.benchmark.client.generate;
 
 import cn.edu.tsinghua.iot.benchmark.client.DataClient;
 import cn.edu.tsinghua.iot.benchmark.distribution.ProbTool;
-import cn.edu.tsinghua.iot.benchmark.entity.Batch;
+import cn.edu.tsinghua.iot.benchmark.entity.Batch.Batch;
 import cn.edu.tsinghua.iot.benchmark.schema.MetaUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
