+++ core/src/main/java/cn/edu/tsinghua/iot/benchmark/client/real/RealDataSetQueryClient.java
@@ -19,7 +19,7 @@
 
 package cn.edu.tsinghua.iot.benchmark.client.real;
 
-import cn.edu.tsinghua.iot.benchmark.entity.Batch;
+import cn.edu.tsinghua.iot.benchmark.entity.Batch.Batch;
 import cn.edu.tsinghua.iot.benchmark.workload.query.impl.VerificationQuery;
 
 import java.util.concurrent.CountDownLatch;
