+++ core/src/main/java/cn/edu/tsinghua/iot/benchmark/extern/CSVDataWriter.java
@@ -19,7 +19,7 @@
 
 package cn.edu.tsinghua.iot.benchmark.extern;
 
-import cn.edu.tsinghua.iot.benchmark.entity.Batch;
+import cn.edu.tsinghua.iot.benchmark.entity.Batch.Batch;
 import cn.edu.tsinghua.iot.benchmark.entity.Record;
 import cn.edu.tsinghua.iot.benchmark.entity.Sensor;
 import cn.edu.tsinghua.iot.benchmark.utils.FileUtils;
