+++ core/src/main/java/cn/edu/tsinghua/iot/benchmark/extern/DataWriter.java
@@ -21,7 +21,7 @@ package cn.edu.tsinghua.iot.benchmark.extern;
 
 import cn.edu.tsinghua.iot.benchmark.conf.Config;
 import cn.edu.tsinghua.iot.benchmark.conf.ConfigDescriptor;
-import cn.edu.tsinghua.iot.benchmark.entity.Batch;
+import cn.edu.tsinghua.iot.benchmark.entity.Batch.Batch;
 
 public abstract class DataWriter {
