+++ core/src/main/java/cn/edu/tsinghua/iot/benchmark/source/CSVDataReader.java
@@ -19,7 +19,7 @@
 
 package cn.edu.tsinghua.iot.benchmark.source;
 
-import cn.edu.tsinghua.iot.benchmark.entity.Batch;
+import cn.edu.tsinghua.iot.benchmark.entity.Batch.Batch;
 import cn.edu.tsinghua.iot.benchmark.entity.Record;
 import cn.edu.tsinghua.iot.benchmark.entity.Sensor;
 import cn.edu.tsinghua.iot.benchmark.schema.MetaDataSchema;
