+++ core/src/main/java/cn/edu/tsinghua/iot/benchmark/source/DataReader.java
@@ -21,7 +21,7 @@ package cn.edu.tsinghua.iot.benchmark.source;
 
 import cn.edu.tsinghua.iot.benchmark.conf.Config;
 import cn.edu.tsinghua.iot.benchmark.conf.ConfigDescriptor;
-import cn.edu.tsinghua.iot.benchmark.entity.Batch;
+import cn.edu.tsinghua.iot.benchmark.entity.Batch.Batch;
 
 import java.util.List;
