+++ core/src/main/java/cn/edu/tsinghua/iot/benchmark/tsdb/fakedb/FakeDB.java
@@ -19,7 +19,7 @@
 
 package cn.edu.tsinghua.iot.benchmark.tsdb.fakedb;
 
-import cn.edu.tsinghua.iot.benchmark.entity.Batch;
+import cn.edu.tsinghua.iot.benchmark.entity.Batch.Batch;
 import cn.edu.tsinghua.iot.benchmark.measurement.Status;
 import cn.edu.tsinghua.iot.benchmark.schema.schemaImpl.DeviceSchema;
 import cn.edu.tsinghua.iot.benchmark.tsdb.IDatabase;
