+++ core/src/main/java/cn/edu/tsinghua/iot/benchmark/workload/interfaces/IQueryWorkLoad.java
@@ -1,8 +1,7 @@
 package cn.edu.tsinghua.iot.benchmark.workload.interfaces;
 
-import cn.edu.tsinghua.iot.benchmark.entity.Batch;
+import cn.edu.tsinghua.iot.benchmark.entity.Batch.Batch;
 import cn.edu.tsinghua.iot.benchmark.exception.WorkloadException;
-import cn.edu.tsinghua.iot.benchmark.workload.query.impl.*;
 import cn.edu.tsinghua.iot.benchmark.workload.query.impl.AggRangeQuery;
 import cn.edu.tsinghua.iot.benchmark.workload.query.impl.AggRangeValueQuery;
 import cn.edu.tsinghua.iot.benchmark.workload.query.impl.AggValueQuery;
