+++ victoriametrics/src/main/java/cn/edu/tsinghua/iot/benchmark/victoriametrics/VictoriaMetrics.java
@@ -20,7 +20,7 @@
 package cn.edu.tsinghua.iot.benchmark.victoriametrics;
 
 import cn.edu.tsinghua.iot.benchmark.conf.Constants;
-import cn.edu.tsinghua.iot.benchmark.entity.Batch;
+import cn.edu.tsinghua.iot.benchmark.entity.Batch.Batch;
 import cn.edu.tsinghua.iot.benchmark.entity.Record;
 import cn.edu.tsinghua.iot.benchmark.entity.Sensor;
 import cn.edu.tsinghua.iot.benchmark.exception.DBConnectException;
